pyproject.toml
[tool.poetry]
name = "notears-analysis"
version = "0.1.0"
description = ""
authors = ["Itsuki Kuwahara <248c117c@stu.kobe-u.ac.jp>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10,<3.11"
causalnex = "^0.12.1"
openpyxl = "^3.1.5"
japanize-matplotlib = "^1.1.3"
geopandas = "^1.0.1"
folium = "^0.19.3"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
